Fix oracleAgent scn NPE (#409)
(cherry picked from commit 1e9131884cc7910751e05e20dc24c357097925c6)

+++ seatunnel-connectors-v2/connector-cdc/connector-cdc-agent-oracle/src/main/java/io/debezium/connector/oracle/Oracle9BridgeStreamingChangeEventSource.java
@@ -84,7 +84,7 @@ public class Oracle9BridgeStreamingChangeEventSource
                     "Start {} from fzsFileNumber={}, scn={}",
                     getClass().getName(),
                     offsetContext.getFzsFileNumber(),
-                    offsetContext.getScn().longValue());
+                    offsetContext.getScn());
             long pollInterval = sourceConfig.getDbzConnectorConfig().getPollInterval().toMillis();
             Oracle9BridgeClient oracle9BridgeClient =
                     Oracle9BridgeClientFactory.getOrCreateStartedSocketClient(
